+++ cw-orch-daemon/src/lib.rs
@@ -27,7 +27,6 @@ mod tx_resp;
 
 pub use self::{builder::*, channel::*, core::*, error::*, state::*, sync::*, tx_resp::*};
 pub use cw_orch_networks::networks;
-pub use network_config::NetworkConfig;
 pub use senders::{query::QuerySender, tx::TxSender, CosmosOptions, Wallet};
 pub use tx_builder::TxBuilder;
 mod cosmos_proto_patches;
