+++ cw-orch-daemon/tests/mock_networks.json
@@ -1,0 +1,29 @@
+{
+  "juno-1": {
+    "chain_id": "joono-1",
+    "gas_denom": "gas_denom",
+    "gas_price": 64.0,
+    "grpc_urls": ["http://juno-grpc.com:123"],
+    "lcd_url": "http://juno-lcd.com:321",
+    "fcd_url": "http://juno-fcd.com:234",
+    "network_info": {
+      "chain_name": "joono",
+      "pub_address_prefix": "joo",
+      "coin_type": 42
+    },
+    "kind": "local"
+  },
+  "neutron-1": { "gas_price": 1.23 },
+  "abstr-1": {
+    "kind": "mainnet",
+    "chain_id": "juno-1",
+    "gas_denom": "ujuno",
+    "gas_price": 0.075,
+    "grpc_urls": ["http://juno-grpc.polkachu.com:12690"],
+    "network_info": {
+      "chain_name": "juno",
+      "pub_address_prefix": "juno",
+      "coin_type": 118
+    }
+  }
+}
